refactor: create product image upload and create product hooks
