+++ finpro-web/src/features/admin/products/utils/product-image-utils.ts
